+++ view/src/components/stock/Items.jsx
@@ -2,13 +2,14 @@ import React from 'react'
 import EMPTY from '@assets/item/empty_img.jpg'
 
 const Items = ({items}) => {
+  console.log(items)
   const baseUrl = import.meta.env.VITE_APP_GATEWAY_URL + '/oauth/file/i/'
   const getFile = (fileNo) => {
     if(fileNo == null) return EMPTY
     return baseUrl + fileNo
   }
   const linkEvent = no => {
-    document.location.href = `/item/${no}`
+    document.location.href = `/stock/${no}`
   }
   return (
     <div className="row g-4 mb-2">
@@ -19,9 +20,10 @@ const Items = ({items}) => {
               <div className="card h-100">
                 <img src={getFile(v.fileNo)} className="card-img-top" />
                 <div className="card-body d-flex flex-column">
-                  <h5 className="card-title mt-2 text-truncate">{v.name}</h5>
+                  <h5 className="card-title mt-2 text-truncate">{v.itemName}</h5>
                   <ul className="list-group list-group-flush mb-3 flex-grow-1">
-                    <li className="list-group-item px-0 py-1"><strong>품목번호:</strong> {v.no}</li>
+                    <li className="list-group-item px-0 py-1"><strong>품목번호:</strong> {v.itemNo}</li>
+                    <li className="list-group-item px-0 py-1"><strong>재고량:</strong> {v.bundle * v.qty}</li>
                   </ul>
                 </div>
               </div>
@@ -33,4 +35,4 @@ const Items = ({items}) => {
   )
 }
 
-export default ItemsList;
+export default Items
