+++ view/src/pages/stock/Detail.jsx
@@ -2,96 +2,8 @@ import { useState, useEffect } from 'react'
 import { useParams } from "react-router-dom"
 import EMPTY from '@assets/item/empty_img.jpg'
 import { POST, GET } from '@utils/Network.js'
-
-const Release = ({releases}) => {
-
-  const linkEvent = no => {
-    document.location.href = `/order/${no}`
-  }
-  
-  return (
-    <>
-    <div className="d-flex mb-2" style={{overflowY: 'auto'}}>
-      <table>
-        <thead>
-          <tr>
-            <th style={{minWidth: '90px'}}>입출고번호</th>
-            <th style={{minWidth: '50px'}}>수량</th>
-            <th style={{minWidth: '140px'}}>출발일시</th>
-            <th style={{minWidth: '140px'}}>도착일시</th>
-          </tr>
-        </thead>
-        <tbody>
-          {releases?.map((v, i) => {
-            console.log(v)
-            return (
-              <tr style={{cursor: 'pointer'}} key={i} onClick={() => linkEvent(v.orderNo)}>
-                <td>{v.no}</td>
-                <td>{v.iqty}</td>
-                <td>{v.depDate}</td>
-                <td>{v.arrDate}</td>
-              </tr>
-            )
-          })}
-          {releases.length == 0 && <tr className='text-center fw-bold fs-6'><td colSpan="4">조회된 입출고 내역이 없습니다.</td></tr>}
-        </tbody>
-      </table>
-    </div>
-    </>
-  )
-}
-
-const Pagination = ({pagination, clickEvent, page, total}) => {
-  
-  const oneStepEvent = type => {
-    if(type === 'p') {
-      clickEvent(page == 0 ? page : page - 1)
-    } 
-    if(type === 'n') {
-      clickEvent(page == total-1 ? page : page + 1)
-    }
-    if(type === 'P') {
-      clickEvent(0)
-    }
-    if(type === 'N') {
-      clickEvent(total-1)
-    }
-  }
-  return (
-    <div>
-      <ul className="pagination justify-content-center">
-        <li className="page-item">
-          <button className="page-link" aria-label="First" onClick={() => oneStepEvent('P')}>
-            <span aria-hidden="true">&laquo;</span>
-          </button>
-        </li>
-        <li className="page-item">
-          <button className="page-link" aria-label="Previous" onClick={() => oneStepEvent('p')}>
-            <span aria-hidden="true">&lsaquo;</span>
-          </button>
-        </li>
-        {
-          pagination?.map((v, i) => {
-            return (
-              <li className="page-item" key={i} ><button className={v.active ? 'page-link active' : 'page-link'}  onClick={()=> clickEvent(i)}>{v.page}</button></li>
-            )
-          })
-        }
-        <li className="page-item">
-          <button className="page-link" aria-label="Next" onClick={() => oneStepEvent('n')}>
-            <span aria-hidden="true">&rsaquo;</span>
-          </button>
-        </li>
-        <li className="page-item">
-          <button className="page-link" aria-label="Last" onClick={() => oneStepEvent('N')}>
-            <span aria-hidden="true">&raquo;</span>
-          </button>
-        </li>
-      </ul>
-    </div>
-    
-  )
-}
+import Release from '@components/stock/Release.jsx'
+import Pagination from '@components/commons/Pagination.jsx'
 
 const Detail = () => {
   const [stock, setStock] = useState({})
@@ -122,7 +34,6 @@ const Detail = () => {
         const totalQty = res.result.stock.bundle * res.result.stock.qty
         const stock = res.result.stock
         setStock({...stock, totalQty})
-        console.log(res.result.release.list)
         setReleases(res.result.release.list)
         const arr = []
         for(let i = 0; i < res.result.release.totalPages; i++) {
@@ -131,7 +42,6 @@ const Detail = () => {
         }
         setPagination(arr)
         setTotal(res.result.release.totalPages)
-        console.log(res.result)
       }
     })
   }
